move file operations to builder-util->fs.ts
